Issue #16 Fix retrieving products from HR VPP got duplicate products.
We retrieve products from VPP in a piecemeal way, because there is a limit on the amount of products you can get in one query.
The client doesn't have full support for pagination, so we adapt our query to limit the number of products per query.
This divides the total termporal extent of the collection into small enough time slots.
But some products have a long enoug period (start/end datetime) that they span several of the time slots we request.
That lead to the producst being received multiple times.

Solution:
We now check if we already have the product and skip it when it is a duplicate.